requirements.txt
pytrends
feedparser
requests
python-dotenv
